Set death save target value if the critical is lower than it.
