+++ scripts/applications/_module.mjs
@@ -1,6 +1,6 @@
-import {BabonusSheet} from "./babonus-sheet.mjs";
-import {BabonusWorkshop} from "./babonus-workshop.mjs";
-import {KeysDialog} from "./keys-dialog.mjs";
+import BabonusSheet from "./babonus-sheet.mjs";
+import BabonusWorkshop from "./babonus-workshop.mjs";
+import KeysDialog from "./keys-dialog.mjs";
 import TokenAura from "./token-aura.mjs";
 
 export default {
