+++ scripts/applications/character-sheet-tab.mjs
@@ -44,7 +44,7 @@ async function _onRenderCharacterSheet2(sheet, [html]) {
   bonuses.all = {label: "BABONUS.Bonuses", key: "all", bonuses: []};
 
   const div = document.createElement("DIV");
-  const isActive = sheet._tabs[0].active === MODULE.ID ? "active" : "";
+  const isActive = (sheet._tabs[0].active === MODULE.ID) ? "active" : "";
   const isEdit = sheet.constructor.MODES.EDIT === sheet._mode;
 
   sheet._filters[MODULE.ID] ??= {name: "", properties: new Set()};
